move kernel acct creation out of contract call func
