+++ src/types.ts
@@ -1,20 +1,29 @@
+import { Chain } from "viem";
+
 export type NetworkProvider = {
   [network: string]: string;
 };
 
 export type AbiAddressPair = {
-  [index: string]: {
-    [key in ContractType]?: {
-      abi: any;
-      address: string;
-    };
-  };
+  abi: any;
+  address: `0x${string}`;
+};
+
+export type ContractToMapping = {
+  [key in ContractType]: AbiAddressPair;
+};
+
+export type ChainInfos = {
+  contracts: ContractToMapping;
+};
+
+export type AllChainInfos = {
+  [key in SupportedNetworks]: ChainInfos;
 };
 
 export enum ContractType {
   DIMO_CREDIT,
   DIMO_REGISTRY,
-  DIMO_TOKEN,
 }
 
 export enum SupportedNetworks {
