+++ src/utils/error.ts
@@ -1,0 +1,7 @@
+export class CustomError extends Error {
+  constructor(message: string) {
+    super(message);
+    Error.captureStackTrace(this, this.constructor);
+    this.name = this.constructor.name;
+  }
+}
