+++ src/web3/actions/mintVehicleWithDeviceDefinition.ts
